Import namaspaces refactoring
1. ProductImport
2. NewsletterUserImport
3. Factories
4. Tests
