+++ tests/AppBundle/Logic/Admin/Import/NewsletterUser/NewsletterUserImportLogicTest.php
@@ -1,14 +1,16 @@
 <?php
 
-namespace Tests\AppBundle\Controller;
+namespace Tests\AppBundle\Logic\Admin\Import\NewsletterUser;
 
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use AppBundle\Logic\Admin\NewsletterUserImportLogic;
 use AppBundle\Factory\Admin\NewsletterUserImportErrorFactory;
 
-class NewsletterUserImportLogicTest extends WebTestCase {
+class ImportLogicTest extends WebTestCase {
+	
 	public function testImport() {
 	}
+	
 	protected function getPreparedRowEntries() {
 		new NewsletterUserImportErrorFactory ( $translator );
 		new NewsletterUserImportLogic ( null, $errorFactory );
